Move extra actions to new file and dummy component to component
