+++ src/System/mod.rs
@@ -1,3 +1,4 @@
+pub mod extra_actions;
 pub mod extract_system_rep;
 pub mod input_enabler;
 pub mod refine;
